rename to AutoConnection, split code to modules and minor code improvements
