+++ src/lib.rs
@@ -9,3 +9,4 @@ pub mod rt;
 pub mod server;
 
 mod error;
+mod util;
