+++ src/util/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) use prepend_async_read::PrependAsyncRead;
+
+mod prepend_async_read;
